Backend/requirements.txt
Django==1.11.11
pytz==2018.3
djangorestframework==3.7.7
django-cors-headers==2.1.0

